+++ resources/views/auto-bulk-form.blade.php
@@ -6,6 +6,6 @@
     /** @var array<int, \Mercurio\Tables\Form\Field\FormField|\Mercurio\Tables\Form\Field\FieldRow> $schema */
 @endphp
 
-<x-tables.bulk-action-form :action="$action" :ids="$ids" :ids-count="$idsCount" :submit-url="$submitUrl">
-    <x-tables.form-renderer :schema="$schema" :model="null"/>
-</x-tables.bulk-action-form>
+<x-tables::bulk-action-form :action="$action" :ids="$ids" :ids-count="$idsCount" :submit-url="$submitUrl">
+    <x-tables::form-renderer :schema="$schema" :model="null"/>
+</x-tables::bulk-action-form>