+++ resources/views/auto-row-form.blade.php
@@ -5,6 +5,6 @@
     /** @var array<int, \Mercurio\Tables\Form\Field\FormField|\Mercurio\Tables\Form\Field\FieldRow> $schema */
 @endphp
 
-<x-tables.row-action-form :action="$action" :submit-url="$submitUrl">
-    <x-tables.form-renderer :schema="$schema" :model="$model"/>
-</x-tables.row-action-form>
+<x-tables::row-action-form :action="$action" :submit-url="$submitUrl">
+    <x-tables::form-renderer :schema="$schema" :model="$model"/>
+</x-tables::row-action-form>